% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_meth.R
\name{analyse_cnv}
\alias{analyse_cnv}
\title{A Function Analyses CNV data}
\usage{
analyse_cnv(gene, cnv_data, cnv_platform, cols, PLOT = FALSE, up_str = 5000,
  dwn_str = 5000, method = NULL, FULL = FALSE)
}
\arguments{
\item{gene}{A vector describing the gene (line of a bed file).}

\item{cnv_data}{A matrix of CN values.}

\item{cnv_platform}{A data frame describing probe positions.}

\item{cols}{A color vectors indexed by by samples names.}

\item{PLOT}{A boolean defining if graphical output must be dispayed on the graphical output.}

\item{up_str}{An integer specifying up stream size (in bp).}

\item{dwn_str}{An integer specifying down stream size (in bp).}

\item{method}{A function used to extract CN score on each region}
}
\description{
This function analyses methylome data.
}

