% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_meth.R
\name{analyse_trscr_cnv}
\alias{analyse_trscr_cnv}
\title{A Function Analyses CNV and Transcriptome data}
\usage{
analyse_trscr_cnv(gene, trscr_res, cnv_res, meth_idx, ctrl_idx, cols, idxes,
  PLOT = TRUE)
}
\arguments{
\item{gene}{A vector describing the gene (line of a bed file).}

\item{trscr_res}{A vector of expression values.}

\item{cnv_res}{A vector of CN values.}

\item{meth_idx}{A vector of sample for which methylom analysis is available.}

\item{ctrl_idx}{A vector of sample used to define threshold}

\item{cols}{A color vectors indexed by by samples names.}

\item{idxes}{A list of index defining sample groups.}

\item{PLOT}{A boolean defining if graphical output must be dispayed on the graphical output.}
}
\description{
This function analyses methylome data.
}

