% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{Description of selected genes}
\format{A data frame with 9 rows and 9 variables:
\describe{
  \item{chrom}{Character string corresponding to chromosome (hg19).}
  \item{chromStart}{Integer corresponding to lowest bound of the gene (hg19).}
  \item{chromEnd}{Integer corresponding  to upperest bound of the gene (hg19).}
  \item{name}{Character string corresponding to the name of the gene.}
  \item{score}{Integer corresponding to nothing (mandatory in bed format).}
  \item{strand}{Character corresponding to the strand of the gene.}
}}
\usage{
genes
}
\description{
A data frame describing selected genes.
}
\keyword{datasets}

