% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_meth.R
\name{plot_meth}
\alias{plot_meth}
\title{A Function that Plots Methylome Analysis Output}
\usage{
plot_meth(meth_analyse, cols, main, alpha.f = 0.1)
}
\arguments{
\item{meth_analyse}{A list, corresponding to analyse_meth function output}

\item{cols}{A color vectors indexed by by samples names.}

\item{main}{A character string specifying the title of the plot.}

\item{alpha.f}{a numeric specifying transparency of convolved signal.}
}
\description{
This function plots methylome analysis output
}

