% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunexp_data}
\alias{sunexp_data}
\title{Beta values of selected CpG sites}
\format{A matrix with 151 rows and 78 columns}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE51954}
}
\usage{
sunexp_data
}
\description{
A matrix of beta values corresponding to 151 probes and 78 samples.
}
\keyword{datasets}

