% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunexp_design}
\alias{sunexp_design}
\title{Genotype of the simulated individuals}
\format{A data frame with 78 rows and 4 variables:
\describe{
  \item{sex}{Character string describing sex of the individual (male or female).}
  \item{age}{Character string describing the age of thne individual.}
  \item{histo}{Character string describing the histological type of the sample (dermis or epidermis).}
  \item{expo}{Character string describing the sun exposure (protected or exposed).}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE51954}
}
\usage{
sunexp_design
}
\description{
A data frame whose rows are markers and columns are individuals. The marker 101 corresponds to the 'all' variable of the 'indiv' data frame, it is the parental allele linked with the phenotype. From this marker 101, genotype is propagated (upstream and downstream) with recombination factor of 5%.
}
\keyword{datasets}

