% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunexp_platform}
\alias{sunexp_platform}
\title{Description of selected CpG sites}
\format{A data frame with 151 rows and 2 variables:
\describe{
  \item{CHR}{Character string corresponding to chromosome (hg19).}
  \item{MAPINFO}{Integer corresponding to position on the chromosome (hg19).}
  \item{UCSC_CpG_Islands_Name}{Character string describing CpG island position (hg19).}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE51954}
}
\usage{
sunexp_platform
}
\description{
A data frame describing CpG positions.
}
\keyword{datasets}

