% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_meth.R
\name{analyse_meth}
\alias{analyse_meth}
\title{A Function Analyses Methylome data}
\usage{
analyse_meth(gene, meth_data, meth_platform, probe_idx,
  pf_chr_colname = "Chromosome", pf_pos_colname = "Start", PLOT = TRUE,
  up_str = 5000, dwn_str = 5000, win_size = 1500, wig_size = 50,
  JUST_PROBE_POS = FALSE, GAUSSIAN_MIXTURE = TRUE, mat_mult = function(A,
  B) {     A \%*\% B })
}
\arguments{
\item{gene}{A vector describing the gene (line of a bed file).}

\item{meth_data}{A matrix of beta values.}

\item{meth_platform}{A data frame describing CpG positions.}

\item{probe_idx}{A vector specifying probes associated to the gene.}

\item{pf_chr_colname}{string matching the name of the column in the platform that contain the chromosome on which we find a probes.}

\item{pf_pos_colname}{string matching the name of the column in the platform that contain the position information of probes.}

\item{PLOT}{A boolean defining if graphical output must be dispayed on the graphical output.}

\item{up_str}{An integer specifying up stream size (in bp).}

\item{dwn_str}{An integer specifying down stream size (in bp).}

\item{win_size}{An integer specifying slidding window size (in bp).}

\item{wig_size}{An integer specifying wiggle size (in bp).}

\item{JUST_PROBE_POS}{A boolean defining if function returns only probes positions.}

\item{GAUSSIAN_MIXTURE}{A boolean specifying if gaussian mixture model is use to convolve signal.}

\item{mat_mult}{A function to multiply matrices.}
}
\value{
A matrix of convolved probes signal around the TSS of the selected gene.
}
\description{
This function analyses methylome data.
}
\examples{
cols = as.numeric(sunexp_design$sex) * 2
names(cols) = rownames(sunexp_design)
gene = genes[1,]
res1 = analyse_meth(gene, sunexp_data, sunexp_platform)
plot_meth(res1, cols)
legend("topright", col=as.numeric(unique(sunexp_design$sex)) * 2, 
       legend=unique(sunexp_design$sex), lty=1)
}

